refactor!: move LimitHands to separate module
This makes `lib.rs` minimal and just a directory to the crate's
submodules.
