+++ src/calc.rs
@@ -1,7 +1,8 @@
+use crate::fu::{calculate_total_fu_value, Fu};
 use crate::hand::error::HandErr;
 use crate::hand::Hand;
+use crate::limit_hand::LimitHands;
 use crate::yaku::Yaku;
-use crate::{calculate_total_fu_value, Fu, LimitHands};
 
 #[derive(Debug, PartialEq)]
 pub enum CalculatorErrors {
